requirements.txt
aprslib==0.6.47
pyyaml
plac
paho-mqtt
